Remove dead perform_kometa_update; move perform_quickstart_update to _qs_update.py
Two cleanups in one branch:

1. Delete perform_kometa_update from _legacy.py. It's a git-based Kometa
   updater that was replaced by perform_kometa_update_zip_only /
   perform_kometa_update_zip_only_at_root (now in _zip_update.py). There
   are zero callers anywhere in the tree. YAGNI.

2. Move perform_quickstart_update from _legacy.py to _qs_update.py where
   it joins the existing Quickstart version/update detection helpers.
   Called from quickstart.py and monkey-patched in tests; both continue
   to resolve via helpers.perform_quickstart_update (star re-export).

Also drops the now-unused 'try: from git import Repo' block from
_legacy.py (GitPython was only used by the deleted perform_kometa_update).

Reduces _legacy.py from 632 to 374 lines (-258). All 1077 tests pass.
