+++ modules/helpers/_legacy.py
@@ -1,5 +1,4 @@
 import os
-import subprocess
 import sys
 import time
 
@@ -10,12 +9,6 @@ from flask import current_app as app
 from flask import has_app_context, has_request_context, session
 
 from modules.helpers._logging import ts_log
-
-try:
-    from git import Repo
-except ImportError:
-    Repo = None  # Prevents errors if GitPython is missing
-
 
 STRING_FIELDS = {"apikey", "token", "username", "password"}
 GITHUB_BASE_URL = "https://raw.githubusercontent.com/Kometa-Team/Kometa"
@@ -259,257 +252,6 @@ def save_to_named_config(yaml_text, config_name, font_refs=None):
 
     # Return POSIX-style filename (used for CLI path like --config config/name_config.yml)
     return latest_path.name
-
-
-def perform_kometa_update(kometa_root, branch="master"):
-    """
-    QS 'master'  -> Kometa 'master'
-    QS != master -> Kometa 'nightly'
-    Deterministic update: fetch -> switch -> reset -> pip upgrade -> requirements
-    """
-    logs, success = [], True
-    try:
-        kometa_root = Path(kometa_root).resolve()
-        is_windows = sys.platform.startswith("win")
-        kometa_branch = "master" if branch == "master" else "nightly"
-        logs.append(f"⚙️ Quickstart branch '{branch}' → using Kometa branch '{kometa_branch}'.")
-
-        if not (kometa_root / ".git").exists():
-            logs.append("❌ Kometa path is not a Git repository (missing .git).")
-            return {"success": False, "log": logs}
-
-        # pick upstream remote if present
-        remotes = subprocess.run(
-            ["git", "remote"],
-            cwd=kometa_root,
-            capture_output=True,
-            text=True,
-            shell=is_windows,
-        ).stdout.split()
-        upstream = "kometa-team" if "kometa-team" in remotes else "origin"
-        logs.append(f"🔗 Using remote: {upstream}")
-
-        # 1) fetch
-        logs.append(f"📥 git fetch {upstream} --prune")
-        p = subprocess.run(
-            ["git", "fetch", upstream, "--prune"],
-            cwd=kometa_root,
-            capture_output=True,
-            text=True,
-            shell=is_windows,
-        )
-        logs.append((p.stdout or "").strip() or "(no output)")
-        if p.returncode != 0:
-            logs.append((p.stderr or "").strip())
-            success = False
-
-        # 2) switch (fallback to checkout)
-        if success:
-            cmd = [
-                "git",
-                "switch",
-                "-C",
-                kometa_branch,
-                "--track",
-                f"{upstream}/{kometa_branch}",
-            ]
-            logs.append(f"🔀 {' '.join(cmd)}")
-            p = subprocess.run(cmd, cwd=kometa_root, capture_output=True, text=True, shell=is_windows)
-            if p.stdout:
-                logs.append(p.stdout.strip())
-            if p.returncode != 0:
-                fallback = [
-                    "git",
-                    "checkout",
-                    "-B",
-                    kometa_branch,
-                    f"{upstream}/{kometa_branch}",
-                ]
-                logs.append(f"🔁 fallback: {' '.join(fallback)}")
-                p = subprocess.run(
-                    fallback,
-                    cwd=kometa_root,
-                    capture_output=True,
-                    text=True,
-                    shell=is_windows,
-                )
-                logs.append((p.stdout or "").strip() or "(no output)")
-                if p.returncode != 0:
-                    logs.append((p.stderr or "").strip())
-                    success = False
-
-        # 3) reset
-        if success:
-            logs.append(f"↩️ git reset --hard {upstream}/{kometa_branch}")
-            p = subprocess.run(
-                ["git", "reset", "--hard", f"{upstream}/{kometa_branch}"],
-                cwd=kometa_root,
-                capture_output=True,
-                text=True,
-                shell=is_windows,
-            )
-            logs.append((p.stdout or "").strip() or "(no output)")
-            if p.returncode != 0:
-                logs.append((p.stderr or "").strip())
-                success = False
-
-        # 4) venv pip upgrade
-        if success:
-            venv_path = kometa_root / "kometa-venv"
-            pip_bin = venv_path / ("Scripts" if is_windows else "bin") / ("pip.exe" if is_windows else "pip")
-            logs.append("\n⬆️ Upgrading pip in Kometa venv...")
-            p = subprocess.run(
-                [str(pip_bin), "install", "--upgrade", "pip"],
-                cwd=kometa_root,
-                capture_output=True,
-                text=True,
-                shell=is_windows,
-            )
-            logs.append((p.stdout or "").strip() or "(no output)")
-            if p.returncode != 0:
-                logs.append((p.stderr or "").strip())
-                success = False
-
-        # 5) install requirements
-        if success:
-            logs.append("\n📦 Installing requirements...")
-            p = subprocess.run(
-                [
-                    str(pip_bin),
-                    "install",
-                    "--no-cache-dir",
-                    "--upgrade",
-                    "-r",
-                    "requirements.txt",
-                ],
-                cwd=kometa_root,
-                capture_output=True,
-                text=True,
-                shell=is_windows,
-            )
-            logs.append((p.stdout or "").strip() or "(no output)")
-            if p.returncode != 0:
-                logs.append((p.stderr or "").strip())
-                success = False
-
-        logs.append("\n✅ Kometa update completed." if success else "\n❌ Kometa update failed.")
-        return {"success": success, "log": logs}
-    except Exception as e:
-        logs.append(f"❌ Exception: {str(e)}")
-        return {"success": False, "log": logs}
-
-
-def perform_quickstart_update(qs_root, branch="master"):
-    """
-    Deterministic Quickstart update (mirrors Kometa updater):
-        - Choose upstream remote: prefer 'kometa-team', else 'origin'
-        - git fetch <upstream> --prune
-        - git switch -C <branch> --track <upstream>/<branch>  (fallback to checkout)
-        - git reset --hard <upstream>/<branch>
-        - python -m pip install --upgrade pip
-        - python -m pip install --no-cache-dir --upgrade -r requirements.txt
-    Returns: {"success": bool, "log": [str, ...]}
-    """
-    logs, success = [], True
-    try:
-        qs_root = Path(qs_root).resolve()
-        is_windows = sys.platform.startswith("win")
-
-        # pick upstream remote (prefer official)
-        remotes_out = subprocess.run(
-            ["git", "remote"],
-            cwd=qs_root,
-            capture_output=True,
-            text=True,
-            shell=is_windows,
-        )
-        remotes = (remotes_out.stdout or "").split()
-        upstream = "kometa-team" if "kometa-team" in remotes else "origin"
-        logs.append(f"🔗 Using Quickstart remote: {upstream}")
-        logs.append(f"⚙️ Target Quickstart branch: {branch}")
-
-        def run(cmd, label=None):
-            if label:
-                logs.append(label)
-            p = subprocess.run(cmd, cwd=qs_root, capture_output=True, text=True, shell=is_windows)
-            out = (p.stdout or "").strip()
-            err = (p.stderr or "").strip()
-            if out:
-                logs.append(out)
-            if p.returncode != 0 and err:
-                logs.append(err)
-            return p
-
-        # 1) fetch (ensure upstream/<branch> exists)
-        p = run(["git", "fetch", upstream, "--prune"], f"📥 git fetch {upstream} --prune")
-        success &= p.returncode == 0
-
-        # 2) switch to branch (fallback to checkout)
-        if success:
-            p = run(
-                ["git", "switch", "-C", branch, "--track", f"{upstream}/{branch}"],
-                f"🔀 git switch -C {branch} --track {upstream}/{branch}",
-            )
-            if p.returncode != 0:
-                p = run(
-                    ["git", "checkout", "-B", branch, f"{upstream}/{branch}"],
-                    f"🔁 fallback: git checkout -B {branch} {upstream}/{branch}",
-                )
-                success &= p.returncode == 0
-
-        # 3) hard reset to upstream tip
-        if success:
-            p = run(
-                ["git", "reset", "--hard", f"{upstream}/{branch}"],
-                f"↩️ git reset --hard {upstream}/{branch}",
-            )
-            success &= p.returncode == 0
-
-        # 4) upgrade pip for this interpreter (QS uses its own Python)
-        if success:
-            logs.append("\n⬆️ Upgrading pip...")
-            p = subprocess.run(
-                [str(Path(sys.executable)), "-m", "pip", "install", "--upgrade", "pip"],
-                cwd=qs_root,
-                capture_output=True,
-                text=True,
-                shell=is_windows,
-            )
-            logs.append((p.stdout or "").strip() or "(no output)")
-            if p.returncode != 0:
-                logs.append((p.stderr or "").strip())
-                success = False
-
-        # 5) install requirements
-        if success:
-            logs.append("\n📦 Installing requirements...")
-            p = subprocess.run(
-                [
-                    str(Path(sys.executable)),
-                    "-m",
-                    "pip",
-                    "install",
-                    "--no-cache-dir",
-                    "--upgrade",
-                    "-r",
-                    "requirements.txt",
-                ],
-                cwd=qs_root,
-                capture_output=True,
-                text=True,
-                shell=is_windows,
-            )
-            logs.append((p.stdout or "").strip() or "(no output)")
-            if p.returncode != 0:
-                logs.append((p.stderr or "").strip())
-                success = False
-
-        logs.append("\n✅ Update completed." if success else "\n❌ Update failed.")
-        return {"success": success, "log": logs}
-
-    except Exception as e:
-        logs.append(f"❌ Exception: {e}")
-        return {"success": False, "log": logs}
 
 
 def get_kometa_root_path() -> Path:
